Moved Flow Sensor Code to Header File
